requirements.txt
numpy
scipy
sounddevice
gTTS
python-dotenv
pypdf
openai
streamlit